chore(mc): Put TopSitesPerfTimer in its own file
